Move EscapeSequenceParser to core
