+++ src/InputHandler.ts
@@ -4,17 +4,18 @@
  * @license MIT
  */
 
-import { IInputHandler, IDcsHandler, IEscapeSequenceParser, IInputHandlingTerminal } from './Types';
+import { IInputHandler, IInputHandlingTerminal } from './Types';
 import { C0, C1 } from 'common/data/EscapeSequences';
 import { CHARSETS, DEFAULT_CHARSET } from 'core/data/Charsets';
 import { wcwidth } from './CharWidth';
-import { EscapeSequenceParser } from './EscapeSequenceParser';
+import { EscapeSequenceParser } from 'core/parser/EscapeSequenceParser';
 import { IDisposable } from 'xterm';
 import { Disposable } from 'common/Lifecycle';
 import { concat } from 'common/TypedArrayUtils';
 import { StringToUtf32, stringFromCodePoint, utf32ToString, Utf8ToUtf32 } from 'core/input/TextDecoder';
 import { CellData, Attributes, FgFlags, BgFlags, AttributeData, NULL_CELL_WIDTH, NULL_CELL_CODE, DEFAULT_ATTR_DATA } from 'core/buffer/BufferLine';
 import { EventEmitter2, IEvent } from 'common/EventEmitter2';
+import { IParsingState, IDcsHandler, IEscapeSequenceParser } from 'core/parser/Types';
 
 /**
  * Map collect to glevel. Used in `selectCharset`.
@@ -277,7 +278,7 @@ export class InputHandler extends Disposable implements IInputHandler {
     /**
      * error handler
      */
-    this._parser.setErrorHandler((state) => {
+    this._parser.setErrorHandler((state: IParsingState) => {
       this._terminal.error('Parsing error: ', state);
       return state;
     });
